requirements.txt
functions-framework==3.*
google-cloud-bigquery
google-auth
google-api-python-client
